requirements.txt
click==6.7
Pillow==5.2.0
